requirements.txt
Flask
Flask-Mail
Flask-SQLAlchemy
Flask-Login
Flask-WTF
email-validator
python-dotenv
pyodbc
